Add precision in documentation that plugin has no requirement
Previously, documentation didn't precise whether it was necessary
or not to install anything on the system running the plugin for it
to work, for example the wsk CLI. This made it unclear for user
how possible it was to use this plugin in a Continuous Deployment
setting.

This commit adds this precision.