Make Docker image name optional
Previously, plugin required that user provide a value for the Docker
image to use as base for the OpenWhisk action. This was unnecessary
friction for user, since many users probably use the default OpenWhisk
Docker image, have no clue what it is, would require to search to find
it, whereas the plugin could use it by default. This made starting up
with OpenWhisk or this plugin less smooth than could be.

This commit improves it by making this parameter optional, and
defaulting to the OpenWhisk skeleton image.