Update example commit hash in documentation
Previously, documentation provided examples with an earlier commit hash.
If user followed documentation, did not provide a Docker image name
and used the commit hash suggested by documentation, they would get a
build failure. This presented the risk of a frustrating start up with
this plugin.

This commit fixes it by replacing the commit hash used in documentation
by one that is compatible with what documentation says.